requirements.txt
asgiref==3.11.0
boto3==1.35.99
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.1.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.4
djangorestframework==3.16.1
gunicorn==22.0.0
idna==3.11
pillow==12.1.0
python-decouple==3.8
requests==2.32.5
sentry-sdk==2.20.0
sqlparse==0.5.4
stripe==14.0.1
typing_extensions==4.15.0
urllib3==2.6.0
whitenoise==6.11.0
